services/auth/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.38.23
botocore==1.38.23
cffi==1.17.1
click==8.1.8
cryptography==45.0.3
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
jmespath==1.0.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.34.2
